+++ docs/changelog.md
@@ -4,7 +4,8 @@ Versions and bullets are arranged chronologically from latest to oldest.
 
 ## v4.0.0 (unreleased)
 
-- \[API\]\[animator\] Add cel() helper
+- \[API\]\[Animator\] Revise API; add reset()
+- \[API\]\[Animator\] Add cel() helper
 - \[build\] Drop explicit NPM silencing
 - \[build\] Colocate tests
 - \[build\] Remove Webpack and NPM workspaces